+++ airflow/contrib/sensors/datadog_sensor.py
@@ -18,8 +18,8 @@
 # under the License.
 from datadog import api
 
-from airflow.contrib.hooks.datadog_hook import DatadogHook
 from airflow.exceptions import AirflowException
+from airflow.providers.datadog.hooks.datadog import DatadogHook
 from airflow.sensors.base_sensor_operator import BaseSensorOperator
 from airflow.utils.decorators import apply_defaults
