+++ scripts/ci/pylint_todo.txt
@@ -8,7 +8,7 @@
 ./airflow/providers/microsoft/azure/hooks/azure_cosmos.py
 ./airflow/providers/microsoft/azure/hooks/azure_data_lake.py
 ./airflow/providers/microsoft/azure/hooks/azure_fileshare.py
-./airflow/contrib/hooks/datadog_hook.py
+./airflow/providers/datadog/hooks/datadog.py
 ./airflow/contrib/hooks/dingding_hook.py
 ./airflow/contrib/hooks/discord_webhook_hook.py
 ./airflow/providers/amazon/aws/hooks/emr.py
