+++ tests/providers/datadog/__init__.py
@@ -16,14 +16,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.datadog.hooks.datadog`."""
-
-import warnings
-
-# pylint: disable=unused-import
-from airflow.providers.datadog.hooks.datadog import DatadogHook  # noqa
-
-warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.datadog.hooks.datadog`.",
-    DeprecationWarning, stacklevel=2
-)
